fix : App, index 파일 분리
