+++ src/App.jsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { useState } from 'react';
+import ClickMe from './ClickMe.jsx';
+import Button from './Button.jsx';
+
+export default function App() {
+  const [counterNumber, setCounterNumber] = useState(0);
+
+  function handlerClickButton({ number }) {
+    setCounterNumber(counterNumber + number);
+  }
+
+  return (
+    <div>
+      <p>Counter</p>
+      <ClickMe
+        counterNumber={counterNumber}
+        handlerClickButton={handlerClickButton}
+      />
+      <br />
+      <Button handlerClickButton={handlerClickButton} />
+    </div>
+  );
+}
